requirements.txt
scikit-build
opencv-python
zbar-py
prestring
tqdm